+++ CiclosEstructurasDeControl/switch.js
@@ -1,0 +1,18 @@
+const dia = 6;
+
+switch(dia){
+    case 0:
+        console.log('Dom');
+        break;
+    case 1:
+        console.log('Lun');
+        break;
+    case 2:
+        console.log('Mar');
+        break;
+    case 3:
+        console.log('Mie');
+        break;
+    default:
+        console.log('Valor Incorrecto');
+}