+++ CHANGELOG.md
@@ -9,10 +9,41 @@
         - Consola
         - Prompt, confirm, alert
 
-
 ### [Fundamentos de JavaScript]
+
     - Inicio[16-06-2022]
         - Tipos de datos Primitivos
         - Palabras reservadas
         - Arreglos
-        - Objetos literales
+        - Objetos literales[22-06-2022]
+        - Más detalles sobre los objetos literales
+        - Funciones básicas y de flecha
+        - Retorno de las funciones
+        - Pro tip: Funciones, argumentos y desestructuración de objetos
+
+### [Ciclos y estructuras de Control]
+
+    -Inicio[23-06-2022]
+        - Valor, referencia y romper la referencia
+        -  If y Else
+        - Laboratorio - Alternativa al if else
+        - Lógica booleana
+
+        - Pro tip: Asignaciones con operadores
+        - Operador condicional ternario
+        - Pro tip: Otros usos del operador ternario
+        - Switch
+        - While y Do While
+        - For - For in - For of
+
+### [Laboratorio 1]
+
+    - Inicio[27-06-2022]
+        - Inicio del proyecto
+        - Nota de actualización
+        - Estilo y estructura
+        - Crear baraja de cartas[28-06-2022]
+        - Pedir carta
+        - Valor de cada carta
+        - Introducción al DOM y su manipulación
+        - Manipulación del DOM - Segunda parte